requirements.txt
atlas-provider-sqlalchemy==0.2.2
click==8.1.7
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2-binary==2.9.9
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.9.0
shellingham==1.5.4
SQLAlchemy==2.0.35
typer==0.12.5
typing_extensions==4.12.2
